content/module7/requirements.txt
transformers
torch
pillow
openai